requirements.txt
# Python 3.12.7
numpy==1.26.4
pandas==2.2.2
requests==2.31.0
python-dotenv==1.2.1
faiss-cpu==1.8.0
pytest==8.0.0
pdfminer.six==20221105
setuptools==80.10.1
natasha==1.6.0
llama-index==0.10.56
llama-index-readers-file==0.1.19

